ExampleDiningCryptographers

Description

n agents dine at a round table [Chaum, 1988]. The dinner may have been paid by their employer (the NSA), or by one of the agents. They execute a protocol to reveal whether one of the agents paid, but without revealing which one. The protocol supplies each pair of adjacent agents with a random coin, which can be observed only by that pair. Each agent announces the result of XORing three Booleans: the two coins observable by her and the status of whether she paid for the dinner. The XOR of all announcements is proven to be equal to the disjunction of whether any agent paid.

Documentation

ag :: Int -> Agent

ag i characterised by its identity i and its nonobservable variables

obs :: Int -> [Var]

observability function, \(ag_i\) observes \(x, p_i, c_i, c_{i+1 \mod n}\),

varx :: Var

boolean variable \(x\)

x :: BExpr

boolean expression x

varc :: Int -> Var

boolean variable \(c_i\) coin flip between agent i and agent i-1

varp :: Int -> Var

boolean variable \(p_i\) agent i paid

allVars :: [Var]

all variables

p :: Int -> BExpr

boolean expression \(p_i\)

c :: Int -> BExpr

boolean expression \(c_i\)

phi :: Formula t

\(\phi\) or \(I\) the constraint on the inital states: "that all agents knows that at most one agent paid"

bexprDC :: BExpr

the boolean expression to be assigned to \(x\) in the algorithm

progDC :: Prog

the DC algorithm as an assignment to \(x\)

n :: Int

the number of agents

noonepaid :: ModalFormula

dontknowwhopaid :: Agent -> ModalFormula

alpha1 :: Formula Modal

alpha2 :: Formula Modal

alpha3 :: Formula Modal

alpha1T :: ModalFormula

alpha2T :: ModalFormula

alpha3T :: ModalFormula

uvar1 :: Int

uvar2 :: Int

uvar3 :: Int

writeForm1 :: IO ()

writing the formulas into a file parsable by Z3py, containing 4 lines line 1 : number of cryptographers line 2 : number of universally quantified variables line 3 : first-order context ϕ line 4 : translation τ(ϕ,α)

writeForm2 :: IO ()

writeForm3 :: IO ()