Logics

Description

The module define the logical languages and everything that is needed inside them * agents and variables * expressions (boolean and integer expressions), * mathcal{L_{DK}}, mathcal{L_{square K}}, mathcal{L}_{FO} formulas * programs inside the box * weakest preconditions, strongest postconditions * substitutions in expressions and formulas

See also the companion paper here https://doi.org/10.48550/arXiv.2206.13841 Maintainer: S.F. Rajaona sfrajaona@gmail.com

Agents and Variables

data Agent

an Agent is determined by its identity and the set of nonobservable variables

Constructors

Agent

Fields

identity :: String

nonobs :: [Var]

data Var

variables are string, we give a domain for integer variables the domain is necessarily only if using SBV, in which quantifications are transformed into conjunctions/disjunctions

Constructors

BVar String

NVar IntDomain String

type IntType = Integer

type IntDomain = [IntType]

varname :: Var -> String

Function varname returns the name of a variable (string)

Expressions (boolean expressions)

data Expr a where

Constructors

B

Fields

:: Bool

-> Expr Bool

Boolean constant

BVal

Fields

:: Var

-> Expr Bool

Boolean variable

BEq :: Expr Bool -> Expr Bool -> Expr Bool

Xor :: [Expr Bool] -> Expr Bool

Or :: [Expr Bool] -> Expr Bool

And :: [Expr Bool] -> Expr Bool

I :: IntType -> Expr Integer

IVal :: Var -> Expr Integer

Add :: Expr Integer -> Expr Integer -> Expr Integer

Mul :: Expr Integer -> Expr Integer -> Expr Integer

Eq :: Expr Integer -> Expr Integer -> Expr Bool

LEq :: Expr Integer -> Expr Integer -> Expr Bool

type BExpr = Expr Bool

type NExpr = Expr Integer

Logical Formulas \(\mathcal{L_{DK}}, \mathcal{L_{\square K}}, \mathcal{L}_{FO} \)

data Formula t where

Constructors

Atom :: BExpr -> Formula t

Neg :: Formula t -> Formula t

Conj :: [Formula t] -> Formula t

Disj :: [Formula t] -> Formula t

Imp :: Formula t -> Formula t -> Formula t

Equiv :: Formula t -> Formula t -> Formula t

K :: Agent -> Formula Modal -> Formula Modal

KV :: Agent -> Var -> Formula Modal

KVe :: Agent -> Var -> Formula Modal

NegKVe :: Agent -> Var -> Formula Modal

Ann :: Formula Modal -> Formula Modal -> Formula Modal

Box :: Prog -> Formula Modal -> Formula Modal

Box' :: Prog -> Formula Modal -> Formula Modal

ForAllB :: Int -> Formula Modal -> Formula Modal

ExistsB :: Int -> Formula Modal -> Formula Modal

ForAllI :: Int -> IntDomain -> Formula Modal -> Formula Modal

ExistsI :: Int -> IntDomain -> Formula Modal -> Formula Modal

kw :: Agent -> ModalFormula -> ModalFormula

data Modal

type ModalFormula = Formula Modal

(∨) :: Formula t -> Formula t -> Formula t

Shotcuts

(∧) :: Formula t -> Formula t -> Formula t

(⇒) :: Formula t -> Formula t -> Formula t

(⇔) :: Formula t -> Formula t -> Formula t

true :: Formula t

false :: Formula t

(⊔) :: Prog -> Prog -> Prog

(≡) :: Expr Integer -> Expr Integer -> Expr Bool

(⨟) :: Prog -> Prog -> Prog

(≤) :: Expr Integer -> Expr Integer -> Expr Bool

(⨰) :: Expr Integer -> Expr Integer -> Expr Integer

(∔) :: Expr Integer -> Expr Integer -> Expr Integer

(⊕) :: Expr Bool -> Expr Bool -> Expr Bool

Mechanism for dealing with quanitifiers

eBVar :: Int -> Var

eBVar constructs an existential boolean variable \(xb_i\) for an integer i

uBVar :: Int -> Var

uBVar constructs an universal boolean variable \(ub_i\) for an integer i

eIVar :: Int -> IntDomain -> Var

eBVar constructs an existential integer variable \(xi_i\) for an integer i

uIVar :: Int -> IntDomain -> Var

uBVar constructs an universal integer variable \(ui_i\) for an integer i

forAllB :: (Var -> Formula Modal) -> Formula Modal

this function chooses an appropriate index \(i\) to build universal quantification \(\exists y_i \ldots\) from of a lambda abstracted ModalFormula.

existsB :: (Var -> Formula Modal) -> Formula Modal

similar to forAllB but with existential

forAllI :: IntDomain -> (Var -> Formula Modal) -> Formula Modal

similar to forAllB but for variables on other domains

existsI :: IntDomain -> (Var -> Formula Modal) -> Formula Modal

similar to existsB but for non-bool domains

maxEBV :: Formula Modal -> Int

maxUBV :: Formula Modal -> Int

maxEIBV :: Formula Modal -> Int

maxUIBV :: Formula Modal -> Int

Programming Language

data Prog

Constructors

Assume ModalFormula

Assert ModalFormula

BAssign Var BExpr

Sequence [Prog]

Nondet [Prog]

Strongest Precondition for comparing with IJCAI'17 paper

sp :: ModalFormula -> Prog -> ModalFormula

Weakest preconditions

wp :: ModalFormula -> Prog -> ModalFormula

Variable Substitution

sub :: Var -> Expr a -> Formula b -> Formula b

sub is the main function, takes the variable, an expression, and the formula to make the substitution

subNExpr :: Var -> NExpr -> NExpr -> NExpr

subNExpr propagates a substitution inside a numerical/integer expression

subBExpr :: Var -> BExpr -> BExpr -> BExpr

subBExpr propagates a substitution inside a boolean expression

subBool :: Var -> BExpr -> Formula a -> Formula a

subBool propagates the substitution made inside a boolean expressions into a formula

subNum :: Var -> NExpr -> Formula a -> Formula a

subNum propagates the substitution inside integer expressions into a formula

Conversion to string

precedence :: Formula t -> Int

toString :: Formula a -> String

toString takes any formula (FOLFormula or ModalFormula) and returns a pretty string.

toPyString :: Formula a -> String

toPythonString is formatted to be parsable by Z3py and CVC5.pythonic.

toPyStringPrec :: Int -> Formula a -> String

toStringPrec :: Int -> Formula a -> String

separateWithCommas :: [String] -> String