Translation

Description

into a \(\mathcal{L}_{FO}\) formula PLEASE Note that the following translation tau by-pass weakest preconditions and does not use public announcement formula, since [β]α = {β?}α

See also the companion paper here https://doi.org/10.48550/arXiv.2206.13841 Maintainer: S.F. Rajaona sfrajaona@gmail.com

Documentation

tau :: ModalFormula -> Formula a -> ModalFormula

The translation function tau takes the constraint formula in FOLFormula and a target formula in ModalFormula

tauSP :: ModalFormula -> Formula a -> ModalFormula

The translation function tau in IJCAI17